+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.3] - 2023-06-28
+* Add keyboard control to `BWidgets::Pattern<T>`
+
+
 ## [1.5.2] - 2023-06-26
 * Add static converters for `BDevices::Keys::KeyType` and key code (`uint32_t`)
 * Add '+'/'-' keyboard controls to dials and sliders